% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_misc_v03.R
\docType{package}
\name{easyPubMed-package}
\alias{easyPubMed}
\alias{easyPubMed-package}
\title{Retrieve and Process Scientific Publication Records from Pubmed}
\description{
Query NCBI Entrez and retrieve PubMed records in XML or TXT format. PubMed records 
can be downloaded and saved as XML or text files. Data integrity is enforced during data download, 
allowing to retrieve and save very large number of records effortlessly. PubMed records can be processed 
to extract publication- and author-specific information.
}
\details{
This software is based on the information 
included in the Entrez Programming Utilities Help manual 
authored by Eric Sayers, PhD and available on the NCBI Bookshelf (NBK25500).
This R library is NOT endorsed, supported, maintained NOR affiliated with NCBI.
}
\examples{
try({
  ## Example 01: retrieve data in XML format, extract info, show
  my_query_string <- "Damiano Fantini[AU]"
  epm <- epm_query(my_query_string)
  epm <- epm_fetch(epm)
  epm <- epm_parse(epm)
  processed_data <- getEPMData(epm)
  utils::head(processed_data)
}, silent = TRUE)

\dontrun{
## Example 02: retrieve data in medline format
try({
  my_query_string <- "Damiano Fantini[AU]"
  epm <- epm_query(my_query_string)
  epm <- epm_fetch(epm, format = 'medline')
  medline_data <- getEPMRaw(epm)
  first_record <- medline_data[[1]] 
  cat(first_record, sep = '\n')
}, silent = TRUE)


## Additional Examples: show easyPubMed Vignette
try({
  library(easyPubMed)
  vignette("easyPubMed_demo")
}, silent = TRUE)

}

}
\references{
\itemize{
  \item Tutorials and Help Webpage: \url{https://www.data-pulse.com/dev_site/easypubmed/}
  \item NCBI PubMed Help Manual: \url{https://pubmed.ncbi.nlm.nih.gov/help/}
  \item Entrez Programming Utilities Help (NBK25500): \url{https://www.ncbi.nlm.nih.gov/books/NBK25500/}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.data-pulse.com/dev_site/easypubmed/}
}

}
\author{
Damiano Fantini \email{damiano.fantini@gmail.com}
}
\keyword{internal}
