% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_classes_and_methods.R
\name{print,easyPubMed-method}
\alias{print,easyPubMed-method}
\title{Print method of the easyPubMed Class.}
\usage{
\S4method{print}{easyPubMed}(x)
}
\arguments{
\item{x}{the `easyPubMed` object being shown.}
}
\description{
Print method of the easyPubMed Class.
}
