% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_classes_and_methods.R
\name{show,easyPubMed-method}
\alias{show,easyPubMed-method}
\title{Show method of the easyPubMed Class.}
\usage{
\S4method{show}{easyPubMed}(object)
}
\arguments{
\item{object}{the `easyPubMed` object being shown.}
}
\description{
Show method of the easyPubMed Class.
}
