% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_all_fx.R
\name{EPM_init_unique_key}
\alias{EPM_init_unique_key}
\title{Generate a Unique Query Key.}
\usage{
EPM_init_unique_key()
}
\value{
string, a 46-char unique key.
}
\description{
Generate a pseudo-random key that uniquely identifies easyPubMed 
objects. The key is a 46-char string
that includes the current date + time and a list of 
randomly selected characters, numbers and special characters.
The unique key is typically saved in the `meta` slot of an
easyPubMed object, and is also written to local files when
records are donwloaded and saved in XML format.
This function takes NO arguments.
}
\examples{
easyPubMed:::EPM_init_unique_key()



}
\references{
\url{https://www.data-pulse.com/dev_site/easypubmed/}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
\keyword{internal}
