% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_all_fx_v01.R
\name{EPM_reference_parse}
\alias{EPM_reference_parse}
\title{Parse and Format References.}
\usage{
EPM_reference_parse(x, max_references = 100, id_type = "pmid")
}
\arguments{
\item{x}{String (character vector of length 1) including an XML Author List
section from a PubMed record.}

\item{max_authors}{Numeric, maximum number of authors to include. See 
details for additional information.}

\item{autofill}{Logical, shall non-missing address information be propagated 
to fill missing address information for other authors 
in the same publication.}
}
\value{
data.frame including one row for each author and n=4 columns:
lastname, forename, address and email.
}
\description{
Extract Reference Information form a raw XML string, typically 
extracted from a PubMed record. If available,
}
\details{
The value of the `max_authors` argument should be tuned to control which 
author information to extract from the input. If 
`max_authors` is set to `0`, no author information are extracted. If 
`max_authors` is set to `-1` (or any negative number), only information
corresponding to the last author are extracted. If `max_authors` is set to 
`+1`, only the first author information are extracted. If `max_authors`
is set to any other positive integer, only information for the 
indicated number of authors is extracted. In this case, information for both 
the first and the last author will be included.
}
\examples{
aff <- paste0('<Author><LastName>Doe</LastName><ForeName>John</ForeName>', 
              '<Affiliation>Univ A</Affiliation></Author>',
              '<Author><LastName>Doe</LastName><ForeName>Jane</ForeName>', 
              '<Affiliation>jane_doe@univ_a.edu</Affiliation></Author>',
              '<Author><LastName>Foo</LastName><ForeName>Bar</ForeName>', 
              '<Affiliation>Univ B</Affiliation></Author>')
easyPubMed:::EPM_auth_parse(aff)



}
\references{
\url{https://www.data-pulse.com/dev_site/easypubmed/}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
\keyword{internal}
