% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_misc.R
\docType{data}
\name{epm_stopwords}
\alias{epm_stopwords}
\title{PubMed Query Stopwords}
\format{
A character vector including all PubMed stopwords tat are typically filtered out from queries.
}
\usage{
data("epm_stopwords")
}
\description{
Collection of 133 Stopwords that can be removed from query strings to improve the 
             accuracy of exact-match PubMed queries.
}
\details{
Number of stopwords included, n=133.
}
\examples{
## Display some contents
data("epm_stopwords")
head(epm_stopwords)
}
\keyword{datasets}
