% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyPubMed_src.R
\name{extract_article_ids}
\alias{extract_article_ids}
\title{Extract Article Ids}
\usage{
extract_article_ids(pubmed_data_list)
}
\arguments{
\item{pubmed_data_list}{List of PubMed records.}
}
\value{
Data Frame including PMID, DOI and PMCID identifiers for each PubMed data included in the input PubMed data list.
}
\description{
Extract PubMed Article Ids (PMID, DOI, PMCID) from a list of PubMed Article Data
}
\references{
\url{https://www.data-pulse.com/dev_site/easypubmed/}
}
\author{
Damiano Fantini \email{damiano.fantini@gmail.com}
}
