% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyPubMed_src.R
\name{fetch_pubmed_data_by_PMID}
\alias{fetch_pubmed_data_by_PMID}
\title{Retrieve PubMed Data in XML or TXT Format based on a list of PMIDs}
\usage{
fetch_pubmed_data_by_PMID(pmids, 
                                 batch = 50, 
                                 format = "xml", 
                                 encoding = "UTF-8", 
                                 verbose = TRUE)
}
\arguments{
\item{pmids}{file name or PMID character vector; this can point to a file including a list of PMIDs 
(for example, a PMID file downloaded from the PubMed website), or a character vector including
a list of PMIDs (example: c("31867428", "31564976", "31564903")).}

\item{batch}{Integer (>10 and <55): number of PMIDs to be retrieved at each iteration.}

\item{format}{Character: element specifying the output format. The following values are allowed: 
c("asn.1", "xml", "medline", "uilist", "abstract"). Note that "uilist" will only retrieve PMIDs.}

\item{encoding}{The encoding of an input/output connection can be specified by name 
(for example, "ASCII", or "UTF-8", in the same way as it would be given to the function base::iconv(). 
See iconv() help page for how to find out more about encodings that can be used on your platform. 
Here, we recommend using "UTF-8".}

\item{verbose}{logical, shall info about the progress of the data download be printed/messaged to console.}
}
\value{
An object (vector) of class "character". If format is set to "xml" (default), a single String including all 
PubMed records (with XML tags embedded) is returned. If a different format is selected, a vector of strings 
is returned, where each row corresponds to a line of the output document.
}
\description{
Retrieve PubMed records from Entrez starting from a list of PumBed IDs (PMIDs). 
Data are downloaded in the XML or TXT format and are retrieved in batches of up to 55 PMIDs
at a time.
}
\details{
Retrieve PubMed records based on a list of PMIDs. 
Records are retrieved from Entrez via the PubMed API efetch function. 
Data will be downloaded on the fly (no files are saved locally).
}
\examples{
try({ 
  ## Example 01: retrieve data in XML format
  library("easyPubMed")
  PMIDs <- c("31506076", "31410434", "31299087", "31552170", 
             "31546103", "31518875", "31518741", "31510010",
             "31509964", "31509747", "31506606", "31503007")
  pmxml <- fetch_pubmed_data_by_PMID(PMIDs)
  # length is 1
  length(pmxml)
  # show an excerpt
  substr(pmxml, 1, 299)
}, silent = TRUE)

}
\references{
\url{https://www.data-pulse.com/dev_site/easypubmed/}
\url{https://www.ncbi.nlm.nih.gov/books/NBK25499/table/chapter4.T._valid_values_of__retmode_and/}
}
\author{
Damiano Fantini \email{damiano.fantini@gmail.com}
}
