% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_classes_and_methods.R
\name{fetchEPMData}
\alias{fetchEPMData}
\alias{fetchEPMData,easyPubMed,list-method}
\title{Method fetchEPMData.}
\usage{
fetchEPMData(x, params)

\S4method{fetchEPMData}{easyPubMed,list}(x, params)
}
\arguments{
\item{x}{an easyPubMed-class object.}

\item{params}{list including parameters to tune the 
record retrieval job. For more info, see 
`?easyPunMed:::EPM_validate_fetch_params`.}
}
\description{
Retrieve PubMed records for an `easyPubMed` object.
}
