% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pubmed_ids2.R
\name{get_pubmed_bib}
\alias{get_pubmed_bib}
\title{Get PubMed Bibliography}
\usage{
get_pubmed_bib(url)
}
\arguments{
\item{url}{String of length 1 specifying the relevant PubMed URL, e.g.,
"https://www.ncbi.nlm.nih.gov/myncbi/<username>/bibliography/public/". For
lists that span multiple pages, the function automatically queries each page
to establish a complete list, thus only the base link should be fed into the
function.}
}
\value{
A data.frame with 11 columns (author list, publication title, journal,
year, full publication date, volume, issue, pages, doi, PubMed ID, and
PubMed Central ID) and number of rows equal to the number of publications
listed on the specified `url`.
}
\description{
Harvests bibliographic content from PubMed static "My Bibliography" pages into
a data frame.
}
