% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_classes_and_methods.R
\name{getEPMData}
\alias{getEPMData}
\alias{getEPMData,easyPubMed-method}
\title{Method getEPMData.}
\usage{
getEPMData(x)

\S4method{getEPMData}{easyPubMed}(x)
}
\arguments{
\item{x}{an object of class `easyPubMed`.}
}
\description{
Retrieve processed data from an `easyPubMed` object.
}
