% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_classes_and_methods.R
\name{getEPMJobList}
\alias{getEPMJobList}
\alias{getEPMJobList,easyPubMed-method}
\title{Method getEPMJobList.}
\usage{
getEPMJobList(x)

\S4method{getEPMJobList}{easyPubMed}(x)
}
\arguments{
\item{x}{an object of class `easyPubMed`.}
}
\description{
Retrieve the list of record retrieval sub-jobs 
from an `easyPubMed` object. 
Record retrieval sub-jobs are stored in a `data.frame` and each
row corresponds to an independent non-overlapping 
PubMed query. This `data.frame` guides the record retrieval process. 
The `data.frame` is obtained from the 
'misc' slot of an `easyPubMed` object.
}
