% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_classes_and_methods.R
\name{getEPMMeta}
\alias{getEPMMeta}
\alias{getEPMMeta,easyPubMed-method}
\title{Method getEPMMeta.}
\usage{
getEPMMeta(x)

\S4method{getEPMMeta}{easyPubMed}(x)
}
\arguments{
\item{x}{an object of class `easyPubMed`.}
}
\description{
Retrieve meta data from an `easyPubMed` object.
}
