% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_classes_and_methods.R
\name{getEPMMisc}
\alias{getEPMMisc}
\alias{getEPMMisc,easyPubMed-method}
\title{Method getEPMMisc.}
\usage{
getEPMMisc(x)

\S4method{getEPMMisc}{easyPubMed}(x)
}
\arguments{
\item{x}{an object of class `easyPubMed`.}
}
\description{
Retrieve miscellaneous information stored in an `easyPubMed` object.
}
