% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_classes_and_methods.R
\name{getEPMQuery}
\alias{getEPMQuery}
\alias{getEPMQuery,easyPubMed-method}
\title{Method getEPMQuery.}
\usage{
getEPMQuery(x)

\S4method{getEPMQuery}{easyPubMed}(x)
}
\arguments{
\item{x}{an object of class `easyPubMed`.}
}
\description{
Retrieve the user-provided query string from an `easyPubMed` object.
}
