% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_classes_and_methods.R
\name{getEPMRaw}
\alias{getEPMRaw}
\alias{getEPMRaw,easyPubMed-method}
\title{Method getEPMRaw.}
\usage{
getEPMRaw(x)

\S4method{getEPMRaw}{easyPubMed}(x)
}
\arguments{
\item{x}{an object of class `easyPubMed`.}
}
\description{
Retrieve the raw PubMed record data stored in an `easyPubMed` object.
}
