% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_classes_and_methods.R
\name{getEPMUilist}
\alias{getEPMUilist}
\alias{getEPMUilist,easyPubMed-method}
\title{Method getEPMUilist.}
\usage{
getEPMUilist(x)

\S4method{getEPMUilist}{easyPubMed}(x)
}
\arguments{
\item{x}{an object of class `easyPubMed`.}
}
\description{
Retrieve the list of unique record identifiers (PMIDs) 
from an `easyPubMed` object.
}
