% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_classes_and_methods.R
\name{parseEPMData}
\alias{parseEPMData}
\alias{parseEPMData,easyPubMed,list-method}
\title{Method parseEPMData.}
\usage{
parseEPMData(x, params)

\S4method{parseEPMData}{easyPubMed,list}(x, params)
}
\arguments{
\item{x}{an easyPubMed-class object}

\item{params}{list including parameters to tune the 
record data parsing job. For more info, see 
`?easyPunMed:::EPM_validate_parse_params`.}
}
\description{
Extract, parse and format information from raw PubMed records 
stored in an `easyPubMed` object.
}
