% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pubmed_ids2.R
\name{scrape_pubmed_element}
\alias{scrape_pubmed_element}
\title{Scrape PubMed Element}
\usage{
scrape_pubmed_element(data, class)
}
\arguments{
\item{data}{the name of the scraped data list (from `rvest::read_html()`).}

\item{class}{the name of the citation element.}
}
\value{
Text string of data from the relevant citation element.
}
\description{
Internal function that harvests relevant data for each element of a citation.
}
