% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_classes_and_methods.R
\name{setEPMData}
\alias{setEPMData}
\alias{setEPMData,easyPubMed,data.frame-method}
\title{Method setEPMData.}
\usage{
setEPMData(x, y)

\S4method{setEPMData}{easyPubMed,data.frame}(x, y)
}
\arguments{
\item{x}{an object of class `easyPubMed`.}

\item{y}{`data.frame` including processed data.}
}
\description{
Attach (or replace) processed data to an `easyPubMed` object.
}
