% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_classes_and_methods.R
\name{setEPMJobList}
\alias{setEPMJobList}
\alias{setEPMJobList,easyPubMed,data.frame-method}
\title{Method setEPMJobList.}
\usage{
setEPMJobList(x, y)

\S4method{setEPMJobList}{easyPubMed,data.frame}(x, y)
}
\arguments{
\item{x}{an object of class `easyPubMed`.}

\item{y}{`data.frame` including the list of 
PubMed record retrieaval sub-jobs.}
}
\description{
Attach (or replace) the list of record retrieval sub-jobs 
to an `easyPubMed` object. 
Record retrieval sub-jobs are stored in a data.frame and each
row corresponds to an independent non-overlapping 
PubMed query. This `data.frame` guides the record retrieval process. 
The `data.frame` is written into the  
`misc` slot of an `easyPubMed` object.
}
