% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_classes_and_methods.R
\name{setEPMMeta}
\alias{setEPMMeta}
\alias{setEPMMeta,easyPubMed,list-method}
\title{Method setEPMMeta.}
\usage{
setEPMMeta(x, y)

\S4method{setEPMMeta}{easyPubMed,list}(x, y)
}
\arguments{
\item{x}{an object of class `easyPubMed`.}

\item{y}{list including meta data information.}
}
\description{
Attach (or replace) meta data to an `easyPubMed` object.
}
