% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_classes_and_methods.R
\name{setEPMMisc}
\alias{setEPMMisc}
\alias{setEPMMisc,easyPubMed,list-method}
\title{Method setEPMMisc.}
\usage{
setEPMMisc(x, y)

\S4method{setEPMMisc}{easyPubMed,list}(x, y)
}
\arguments{
\item{x}{an object of class `easyPubMed`.}

\item{y}{list including miscellaneous data and information.}
}
\description{
Attach (or replace)  miscellaneous information 
to an `easyPubMed` object.
}
