% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_classes_and_methods.R
\name{setEPMQuery}
\alias{setEPMQuery}
\alias{setEPMQuery,easyPubMed,character-method}
\title{Method setEPMQuery.}
\usage{
setEPMQuery(x, y)

\S4method{setEPMQuery}{easyPubMed,character}(x, y)
}
\arguments{
\item{x}{an object of class `easyPubMed`.}

\item{y}{string (character vector of length 1) corresponding to
a PubMed query string.}
}
\description{
Attach (or replace) a user-provided query string 
to an `easyPubMed` object.
}
