% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_classes_and_methods.R
\name{setEPMRaw}
\alias{setEPMRaw}
\alias{setEPMRaw,easyPubMed,list-method}
\title{Method setEPMRaw.}
\usage{
setEPMRaw(x, y)

\S4method{setEPMRaw}{easyPubMed,list}(x, y)
}
\arguments{
\item{x}{an object of class `easyPubMed`.}

\item{y}{list of PubMed records (raw data).}
}
\description{
Attach (or replace) raw PubMed record data 
to an `easyPubMed` object.
}
