% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_classes_and_methods.R
\name{setEPMUilist}
\alias{setEPMUilist}
\alias{setEPMUilist,easyPubMed,list-method}
\title{Method setEPMUilist.}
\usage{
setEPMUilist(x, y)

\S4method{setEPMUilist}{easyPubMed,list}(x, y)
}
\arguments{
\item{x}{an object of class `easyPubMed`.}

\item{y}{list of unique PubMed record identifiers (PMIDs).}
}
\description{
Attach (or replace)  the list of unique record identifiers (PMIDs) 
to an `easyPubMed` object.
}
