% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyPubMed_scr.R
\docType{data}
\name{PubMed_stopwords}
\alias{PubMed_stopwords}
\title{PubMed Records about Bladder Research from Northwestern University}
\format{A character vector including all PubMed stopwords tat are typically filtered out from queries.}
\usage{
data("PubMed_stopwords")
}
\description{
This dataset includes a collection of 87 PubMed Records of scientific publications about Bladder 
biology and pathology, published by clinical and research groups from Northwestern 
University (Chicago, IL), between 2016 and 2018.
}
\details{
Number of stopwords included, n=133.
}
\examples{
## Display some contents
data("PubMed_stopwords")
head(PubMed_stopwords)
}
\keyword{datasets}
